+++ client/scripts/db/dw.js
@@ -85,7 +85,13 @@ module.exports = {
   ],
   skipWmcRomCheck: true,
   initialise: {
-    closedSwitches: [ 22, 25, 26, 27, 82 ],
+    closedSwitches: [
+      22, 25, 26, 27,
+      82,
+      //OPTO SWITCHES
+      31, 32, 33, 71, 72, 73, 74, 75, 76, 77,
+      'F2', 'F4', 'F6', 'F8',
+    ],
     initialAction: [
       {
         delayMs: 1000,